fix(core): drop (int)size truncation in flat argsort/argpartition/cum-scan output shapes
The flat (axis=null) branches of np.argsort, np.argpartition and the
cumsum/cumprod empty-result path built their 1-D output shape with
new Shape((int)a.size) / Shape.Vector((int)shape.size). For an array
whose total element count exceeds int.MaxValue the (int) cast overflows
(e.g. 3e9 -> negative), so the output shape was garbage/negative even
though the adjacent reshape(a.size) already used the long overload.

Storage is long-capable (UnmanagedMemoryBlock<T>.Count / UnmanagedStorage.Count
are long) and Shape carries long dims/strides/size, so these casts were
gratuitous: Shape.Vector(long) and Shape(params long[]) already exist.
Switched all five sites to Shape.Vector(a.size)/Shape.Vector(shape.size).

The cum-scan sites are only reached with size==0 (harmless today) but are
corrected for consistency; the argsort/argpartition sites are reachable
non-empty flat paths and were genuine >int.MaxValue truncation bugs.

Not touched (verified correct): Shape's explicit (int) operator and
UnmanagedStorage.ToArray()/GetSpan() guard with OverflowException/
InvalidOperationException before casting; Char->string, tobytes, nditer<T>
Span, ArrayPool.Rent and managed T[] factories are inherently int-bounded
framework APIs (several already checked); QuantileEngine flat n=(int)size
is a genuine >int.MaxValue limit but threads int n through an IL kernel,
left for a dedicated change.

+++ src/NumSharp.Core/Backends/Default/Math/Reduction/Default.Reduction.CumAdd.cs
@@ -47,7 +47,7 @@ namespace NumSharp.Backends
             // (0,)); with an axis the shape is preserved.
             if (shape.IsEmpty || shape.size == 0)
                 return new NDArray(retTypeCode,
-                    axis_ == null ? Shape.Vector((int)shape.size) : new Shape(shape.dimensions), false);
+                    axis_ == null ? Shape.Vector(shape.size) : new Shape(shape.dimensions), false);
 
             // 0-d scalar or single-element 1-D: cumsum is the value itself, promoted to the
             // accumulator dtype and shaped 1-D — cumsum NEVER returns 0-d (NumPy: cumsum(0-d) -> (1,),
